Fix refresh migrations. Fixes #8
