+++ package.json
@@ -16,6 +16,7 @@
     "entity",
     "framework",
     "migrations",
+    "entity framework",
     ".net",
     "dotnet",
     "core",
