+++ src/actions/AddMigrationAction.ts
@@ -1,15 +1,21 @@
 import * as vscode from 'vscode';
+import { CommandProvider } from '../commands/CommandProvider';
+import { RefreshTreeCommand } from '../commands/RefreshTreeCommand';
 import { getCommandsConfig } from '../config/config';
 
 import type { TerminalProvider } from '../terminal/TerminalProvider';
+import {
+  dbContextsCache,
+  DbContextTreeItem,
+} from '../treeView/DbContextTreeItem';
 import { TerminalAction } from './TerminalAction';
 
 export class AddMigrationAction extends TerminalAction {
   constructor(
     terminalProvider: TerminalProvider,
-    workspaceRoot: string,
-    dbContext: string,
-    project: string,
+    private readonly workspaceRoot: string,
+    private readonly dbContext: string,
+    private readonly project: string,
   ) {
     super(
       terminalProvider,
@@ -31,9 +37,19 @@ export class AddMigrationAction extends TerminalAction {
     if (!migrationName) {
       return;
     }
-    return super.run({
+    await super.run({
       ...this.params,
       migrationName,
     });
+    const cacheId = DbContextTreeItem.getCacheId(
+      this.workspaceRoot,
+      this.project,
+      this.dbContext,
+    );
+    dbContextsCache.clear(cacheId);
+    await vscode.commands.executeCommand(
+      CommandProvider.getCommandName(RefreshTreeCommand.commandName),
+      false,
+    );
   }
 }
